Fixed Issue #53
As long as the friend code is valid, if a logged in user decides to unfriend a user in his/her friend's list, they can do so successfully. However, that user being unfriended will automatically unfriend the user that is unfriending him/her. All in all, both users will be removed from their respective 'friends' array.