Finished Issue #59
Now when searching up leagues, they will only display if they are public, contain whatever is being searched for in the title, and the end date of the league must be greater than or equal to the current date. This means that if today's date is 11/15/22 and the league being searched for ends in 11/15/22 it will display in the results.