In progress Issue #60
The problem is that when a user creates a league and selects for example 11/20/22 for the start date and 11/21/22 for the end date, it gets stored in UTC format. The UTC format for today's date 11/17/22 is 2022-11-18T06:19:54.661Z (notice that the day changes from 17 to 18). We were comparing that UTC start date with the current UTC date, which explains leagues being passed earlier than supposed to. If instead we compare the stored UTC start date with the actual current date not in UTC format, it should work properly. Still needs testing to ensure, but I'm pretty sure this fixes the problem.