Updated 'join' in leagues.js
A user can now only join a league if the current date(month/day/year) is less than the league start date(month/day/year). In other words if the current date is 11/20/2022 and the league start date is 11/20/2022, user won't be able to join. If the current date is 11/20/2022 and the league start date is 11/19/2022, user won't be able to join. If the current date is 11/20/2022 and the league start date is 11/21/2022, user will be able to join. They must join at least 1 day in advance, can not join the same day the league starts.