Fixed Issue #122
Users who are not joined in a league will not be able to post comments or reply to comments in a league. They will only have access to that functionality if they join the league. This will also prevent the issue with a league appearing blank for other users if a user who isn't in that league leaves a comment or reply in that league and then deletes their account.